Removed most comments, put back summer flowers, and sped up duck
I removed most of the code I had commented out. I also uncommented the for loop that randomly placed flowers in summer. Before, the extra time spent on that for loop caused the duck to slow down its movement in the summer. However I added a line in the summer function that multiplies the duck's x velocity so as to compensate for it. Now, the duck appears to move at the same speed for all seasons where it is moving.